Replaced SiteContextManager lock factory with striped locks (#818)
* Added support for the application cache

* Fixes to cache REST controller so the same API can be used for any type of cache

* Minor enhancements to S3 serverless related classes

* CR recommendations

* CR recommendations

* Revert "CR recommendations"

This reverts commit f1156d0f7c6da5d58896869f33e30ffcb208b535.

* Fixing S3ContentStoreAdapter

* Replaced SiteContextManager lock factory with striped locks

* CR recommendations

* CR recommendations

---------

+++ src/main/java/org/craftercms/engine/service/context/SiteContextManager.java
@@ -15,11 +15,10 @@
  */
 package org.craftercms.engine.service.context;
 
+import com.google.common.util.concurrent.Striped;
 import io.methvin.watcher.DirectoryWatcher;
 import io.methvin.watcher.hashing.FileHasher;
 import org.apache.commons.collections4.CollectionUtils;
-import org.craftercms.commons.concurrent.locks.KeyBasedLockFactory;
-import org.craftercms.commons.concurrent.locks.WeakKeyBasedReentrantLockFactory;
 import org.craftercms.commons.entitlements.exception.EntitlementException;
 import org.craftercms.commons.entitlements.model.EntitlementType;
 import org.craftercms.commons.entitlements.validator.EntitlementValidator;
@@ -41,7 +40,6 @@ import java.util.*;
 import java.util.concurrent.*;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.locks.Lock;
-import java.util.concurrent.locks.ReentrantLock;
 import java.util.function.Consumer;
 import java.util.stream.Collectors;
 
@@ -58,7 +56,7 @@ public class SiteContextManager implements ApplicationContextAware, DisposableBe
     private static final Logger logger = LoggerFactory.getLogger(SiteContextManager.class);
 
     protected ApplicationContext applicationContext;
-    protected KeyBasedLockFactory<ReentrantLock> siteLockFactory;
+    protected Striped<Lock> siteLocks;
     protected Map<String, SiteContext> contextRegistry;
     protected SiteContextFactory contextFactory;
     protected SiteContextFactory fallbackContextFactory;
@@ -133,11 +131,11 @@ public class SiteContextManager implements ApplicationContextAware, DisposableBe
                               Executor jobThreadPoolExecutor, final String defaultSiteName, final int contextBuildRetryMaxCount,
                               final long contextBuildRetryWaitTimeBase, final int contextBuildRetryWaitTimeMultiplier,
                               final boolean modePreview, final String[] watcherPaths, final String[] watcherIgnorePaths,
-                              final int watcherCounterLimit, final int watcherIntervalPeriod) {
-        siteLockFactory = new WeakKeyBasedReentrantLockFactory();
+                              final int watcherCounterLimit, final int watcherIntervalPeriod, final int siteLocksStripeCount) {
+        siteLocks = Striped.lazyWeakLock(siteLocksStripeCount);
         contextRegistry = new ConcurrentHashMap<>();
         directoryWatcherRegistry = new ConcurrentHashMap<>();
-        directoryWatcherLastProcessedHash = new HashMap<>();
+        directoryWatcherLastProcessedHash = new ConcurrentHashMap<>();
         directoryWatcherCounter = new ConcurrentHashMap<>();
         directoryWatcherExecutor = new ConcurrentHashMap<>();
 
@@ -255,7 +253,7 @@ public class SiteContextManager implements ApplicationContextAware, DisposableBe
                                 // This prevents multiple events for batch files change with same modified date such as from a git pull
                                 String lastProcessedHash = directoryWatcherLastProcessedHash.get(siteName);
                                 if (lastProcessedHash == null || event.hash() == null || !lastProcessedHash.equals(hashValue)) {
-                                    Lock siteLock = siteLockFactory.getLock(siteName);
+                                    Lock siteLock = siteLocks.get(siteName);
                                     siteLock.lock();
                                     try {
                                         if (event.hash() != null) {
@@ -283,13 +281,18 @@ public class SiteContextManager implements ApplicationContextAware, DisposableBe
 
             // Remove old watcher before register a new one
             if (directoryWatcherRegistry.get(siteName) != null) {
-                Lock siteLock = siteLockFactory.getLock(siteName);
+                DirectoryWatcher oldWatcher;
+
+                Lock siteLock = siteLocks.get(siteName);
                 siteLock.lock();
                 try {
-                    DirectoryWatcher oldWatcher = directoryWatcherRegistry.remove(siteName);
-                    oldWatcher.close();
+                    oldWatcher = directoryWatcherRegistry.remove(siteName);
                 } finally {
                     siteLock.unlock();
+                }
+
+                if (oldWatcher != null) {
+                    oldWatcher.close();
                 }
             }
             directoryWatcherRegistry.put(siteName, watcher);
@@ -312,11 +315,13 @@ public class SiteContextManager implements ApplicationContextAware, DisposableBe
     public void registerPreviewRebuildTask(String siteName, boolean isFallback) {
         // Remove old executor then register a new one
         if (directoryWatcherExecutor.get(siteName) != null) {
-            Lock siteLock = siteLockFactory.getLock(siteName);
+            Lock siteLock = siteLocks.get(siteName);
             siteLock.lock();
             try {
                 ScheduledExecutorService oldExecutor = directoryWatcherExecutor.remove(siteName);
-                oldExecutor.shutdown();
+                if (oldExecutor != null) {
+                    oldExecutor.shutdown();
+                }
             } finally {
                 siteLock.unlock();
             }
@@ -404,29 +409,42 @@ public class SiteContextManager implements ApplicationContextAware, DisposableBe
             logger.info("<Destroying site context: '{}'>", siteName);
             logger.info("==================================================");
 
-            Lock siteLock = siteLockFactory.getLock(siteName);
+            DirectoryWatcher watcher;
+            ScheduledExecutorService executor;
+
+            Lock siteLock = siteLocks.get(siteName);
             siteLock.lock();
             try {
-                if (directoryWatcherRegistry.get(siteName) != null) {
-                    DirectoryWatcher watcher = directoryWatcherRegistry.remove(siteName);
+                watcher = directoryWatcherRegistry.remove(siteName);
+                executor = directoryWatcherExecutor.remove(siteName);
+                iter.remove();
+            } finally {
+                siteLock.unlock();
+            }
+
+            if (watcher != null) {
+                try {
                     watcher.close();
+                } catch (Exception e) {
+                    logger.warn("Error while closing directory watcher for site '{}'", siteName, e);
                 }
-                if (directoryWatcherExecutor.get(siteName) != null) {
-                    ScheduledExecutorService executor = directoryWatcherExecutor.remove(siteName);
-                    executor.shutdown();
+            }
+            if (executor != null) {
+                try {
+                    executor.shutdownNow();
+                } catch (Exception e) {
+                    logger.warn("Error while shutting down directory watcher executor for site '{}'", siteName, e);
                 }
+            }
+            try {
                 destroyContext(siteContext);
             } catch (Exception e) {
                 logger.error("Error destroying site context for site '{}'", siteName, e);
-            } finally {
-                siteLock.unlock();
             }
 
             logger.info("==================================================");
             logger.info("</Destroying site context: '{}'>", siteName);
             logger.info("==================================================");
-
-            iter.remove();
         }
 
         logger.info("==================================================");
@@ -450,7 +468,7 @@ public class SiteContextManager implements ApplicationContextAware, DisposableBe
                 return null;
             }
 
-            Lock siteLock = siteLockFactory.getLock(siteName);
+            Lock siteLock = siteLocks.get(siteName);
             siteLock.lock();
             try {
                 // Double check locking, in case the context has been created already by another thread
@@ -580,37 +598,43 @@ public class SiteContextManager implements ApplicationContextAware, DisposableBe
      * @param siteName the site name of the context to destroy
      */
     protected void removeSiteContext(String siteName) {
-        SiteContext siteContext;
-
         logger.info("==================================================");
         logger.info("<Removing site context: '{}'>", siteName);
         logger.info("==================================================");
 
-        Lock siteLock = siteLockFactory.getLock(siteName);
+        SiteContext siteContext;
+        DirectoryWatcher watcher;
+        ScheduledExecutorService executor;
+
+        Lock siteLock = siteLocks.get(siteName);
         siteLock.lock();
         try {
-            if (directoryWatcherRegistry.get(siteName) != null) {
-                try {
-                    DirectoryWatcher watcher = directoryWatcherRegistry.remove(siteName);
-                    watcher.close();
-                } catch (IOException e) {
-                    logger.warn("Error while removing directory watcher register for site '{}'", siteName, e);
-                }
-            }
-
-            if (directoryWatcherExecutor.get(siteName) != null) {
-                ScheduledExecutorService executor = directoryWatcherExecutor.remove(siteName);
-                executor.shutdown();
-            }
-
+            watcher = directoryWatcherRegistry.remove(siteName);
+            executor = directoryWatcherExecutor.remove(siteName);
             siteContext = contextRegistry.remove(siteName);
         } finally {
             siteLock.unlock();
         }
 
+        if (watcher != null) {
+            try {
+                watcher.close();
+            } catch (Exception e) {
+                logger.warn("Error while closing directory watcher for site '{}'", siteName, e);
+            }
+        }
+        if (executor != null) {
+            try {
+                executor.shutdownNow();
+            } catch (Exception e) {
+                logger.warn("Error while shutting down directory watcher executor for site '{}'", siteName, e);
+            }
+        }
         if (siteContext != null) {
             try {
                 destroyContext(siteContext);
+            } catch (Exception e) {
+                logger.error("Error destroying site context for site '{}'", siteName, e);
             } finally {
                 applicationContext.publishEvent(new SiteContextRemovedEvent(siteContext));
             }
@@ -641,7 +665,7 @@ public class SiteContextManager implements ApplicationContextAware, DisposableBe
     }
 
     protected SiteContext rebuildContext(String siteName, boolean fallback) {
-        Lock siteLock = siteLockFactory.getLock(siteName);
+        Lock siteLock = siteLocks.get(siteName);
         siteLock.lock();
         try {
             logger.info("==================================================");
@@ -651,7 +675,9 @@ public class SiteContextManager implements ApplicationContextAware, DisposableBe
             SiteContext oldSiteContext = contextRegistry.get(siteName);
             SiteContext newContext = createContext(siteName, fallback);
 
-            oldSiteContext.destroy();
+            if (oldSiteContext != null) {
+                oldSiteContext.destroy();
+            }
 
             logger.info("==================================================");
             logger.info("</Rebuilding site context: '{}'>", siteName);
